+++ app/assets/javascripts/clickable_row.js
@@ -1,0 +1,8 @@
+jQuery(document).ready(function($) {
+  $(".clickable-row").click(function() {
+    var url = $(this).data("href");
+    if (url) {
+      window.location = url;
+    }
+  });
+});
